ejercicioSOA/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
python-dotenv
passlib[bcrypt]
